+++ src/compas/datastructures/network/__init__.py
@@ -11,7 +11,7 @@ from .combinatorics import *  # noqa: F401 F403
 from .complementarity import *  # noqa: F401 F403
 from .duality import *  # noqa: F401 F403
 from .explode import *  # noqa: F401 F403
-from .parallelisation import *  # noqa: F401 F403
+# from .parallelisation import *  # noqa: F401 F403
 
 if not compas.IPY:
     from .planarity_ import *  # noqa: F401 F403