+++ src/compas/datastructures/network/planarity_.py
@@ -11,6 +11,7 @@ import planarity
 from compas.geometry import angle_vectors_xy
 from compas.geometry import is_intersection_segment_segment_xy
 from compas.geometry import is_ccw_xy
+from compas.geometry import subtract_vectors_xy
 
 
 __all__ = [
@@ -48,16 +49,16 @@ def network_is_crossed(network):
             if u1 == u2 or v1 == v2 or u1 == v2 or u2 == v1:
                 continue
             else:
-                a = network.vertex[u1]['x'], network.vertex[u1]['y']
-                b = network.vertex[v1]['x'], network.vertex[v1]['y']
-                c = network.vertex[u2]['x'], network.vertex[u2]['y']
-                d = network.vertex[v2]['x'], network.vertex[v2]['y']
+                a = network.node_attributes(u1, 'xy')
+                b = network.node_attributes(v1, 'xy')
+                c = network.node_attributes(u2, 'xy')
+                d = network.node_attributes(v2, 'xy')
                 if is_intersection_segment_segment_xy((a, b), (c, d)):
                     return True
     return False
 
 
-def _network_are_edges_crossed(edges, vertices):
+def _are_edges_crossed(edges, vertices):
     for u1, v1 in edges:
         for u2, v2 in edges:
             if u1 == u2 or v1 == v2 or u1 == v2 or u2 == v1:
@@ -121,10 +122,10 @@ def network_find_crossings(network):
                     continue
                 if ((u2, v2), (u1, v1)) in crossings:
                     continue
-                a = network.vertex_coordinates(u1, 'xy')
-                b = network.vertex_coordinates(v1, 'xy')
-                c = network.vertex_coordinates(u2, 'xy')
-                d = network.vertex_coordinates(v2, 'xy')
+                a = network.node_attributes(u1, 'xy')
+                b = network.node_attributes(v1, 'xy')
+                c = network.node_attributes(u2, 'xy')
+                d = network.node_attributes(v2, 'xy')
                 if is_intersection_segment_segment_xy((a, b), (c, d)):
                     crossings.add(((u1, v1), (u2, v2)))
     return list(crossings)
@@ -146,11 +147,11 @@ def network_is_xy(network):
 
     """
     z = None
-    for key in network.vertices():
+    for key in network.nodes():
         if z is None:
-            z = network.vertex[key].get('z', 0.0)
+            z = network.node_attribute(key, 'z') or 0.0
         else:
-            if z != network.vertex[key].get('z', 0.0):
+            if z != network.node_attribute(key, 'z') or 0.0:
                 return False
     return True
 
@@ -188,35 +189,7 @@ def network_is_planar(network):
 
     Examples
     --------
-    .. plot::
-        :include-source:
-
-        import compas
-
-        from compas.datastructures import Network
-        from compas.datastructures import network_is_planar
-        from compas.datastructures import network_find_crossings
-        from compas_plotters import NetworkPlotter
-
-        network = Network.from_obj(compas.get('lines.obj'))
-
-        network.add_edge(6, 15)
-
-        if not network_is_planar(network):
-            print('here')
-            crossings = network_find_crossings(network)
-        else:
-            crossings = []
-
-        print(crossings)
-
-        plotter = NetworkPlotter(network)
-
-        plotter.draw_vertices(radius=0.15, text={key: key for key in network.vertices()})
-        plotter.draw_edges(color={edge: '#ff0000' for edges in crossings for edge in edges})
-
-        plotter.show()
-
+    >>>
     """
     return planarity.is_planar(list(network.edges()))
 
@@ -270,36 +243,7 @@ def network_embed_in_plane(network, fix=None, straightline=True):
 
     Examples
     --------
-    .. plot::
-        :include-source:
-
-        import compas
-
-        from compas.datastructures import Network
-        from compas.datastructures import network_embed_in_plane
-        from compas_plotters import NetworkPlotter
-
-        network = Network.from_obj(compas.get('fink.obj'))
-
-        embedding = network.copy()
-
-        fix = (9, 6)
-
-        if network_embed_in_plane(embedding, fix=fix):
-
-            plotter = NetworkPlotter(embedding)
-
-            plotter.draw_lines([{'start': network.vertex_coordinates(u, 'xy'),
-                                  'end': network.vertex_coordinates(v, 'xy'),
-                                  'color': '#cccccc'} for u, v in network.edges()])
-
-            plotter.draw_vertices(radius=0.3,
-                                  text={key: key for key in embedding.vertices()},
-                                  facecolor={key: '#ff0000' for key in fix})
-
-            plotter.draw_edges()
-            plotter.show()
-
+    >>>
     """
     try:
         import networkx as nx
@@ -307,8 +251,8 @@ def network_embed_in_plane(network, fix=None, straightline=True):
         print("NetworkX is not installed. Get NetworkX at https://networkx.github.io/.")
         raise
 
-    x = network.get_vertices_attribute('x')
-    y = network.get_vertices_attribute('y')
+    x = network.vertices_attribute('x')
+    y = network.vertices_attribute('y')
     xmin, xmax = min(x), max(x)
     ymin, ymax = min(y), max(y)
     xspan = xmax - xmin
@@ -321,11 +265,8 @@ def network_embed_in_plane(network, fix=None, straightline=True):
     count = 100
     while count:
         graph = nx.Graph(edges)
-        pos = nx.spring_layout(graph,
-                               2,
-                               iterations=100,
-                               scale=max(xspan, yspan))
-        if not _network_are_edges_crossed(edges, pos):
+        pos = nx.spring_layout(graph, 2, iterations=100, scale=max(xspan, yspan))
+        if not _are_edges_crossed(edges, pos):
             is_embedded = True
             break
         count -= 1
@@ -335,11 +276,11 @@ def network_embed_in_plane(network, fix=None, straightline=True):
 
     if fix:
         a, b = fix
-        p0 = network.vertex_coordinates(a, 'xy')
-        p1 = network.vertex_coordinates(b, 'xy')
+        p0 = network.node_attributes(a, 'xy')
+        p1 = network.node_attributes(b, 'xy')
         p2 = pos[b]
-        vec0 = [network.vertex[b][axis] - network.vertex[a][axis] for axis in 'xy']
-        vec1 = [pos[b][axis] - pos[a][axis] for axis in (0, 1)]
+        vec0 = subtract_vectors_xy(p1, p0)
+        vec1 = subtract_vectors_xy(pos[b], pos[a])
         # rotate
         a = angle_vectors_xy(vec0, vec1)
         if is_ccw_xy(p0, p1, p2):
@@ -358,7 +299,7 @@ def network_embed_in_plane(network, fix=None, straightline=True):
             pos[key][0] *= scale
             pos[key][1] *= scale
         # translate
-        t = network.vertex[fix[0]]['x'] - pos[fix[0]][0], network.vertex[fix[0]]['y'] - pos[fix[0]][1]
+        t = subtract_vectors_xy(p0, pos[a])
         for key in pos:
             pos[key][0] += t[0]
             pos[key][1] += t[1]
@@ -366,8 +307,7 @@ def network_embed_in_plane(network, fix=None, straightline=True):
     # update network vertex coordinates
     for key in network.vertices():
         if key in pos:
-            network.vertex[key]['x'] = pos[key][0]
-            network.vertex[key]['y'] = pos[key][1]
+            network.node_attributes(key, 'xy', pos[key])
 
     return True
 
@@ -394,9 +334,10 @@ if __name__ == '__main__':
 
     print(crossings)
 
-    plotter = NetworkPlotter(network)
+    plotter = NetworkPlotter(network, figsize=(8, 5))
+    plotter.defaults['node.fontsize'] = 6
 
-    plotter.draw_vertices(radius=0.15, text={key: key for key in network.vertices()})
+    plotter.draw_nodes(radius=0.15, text={key: key for key in network.nodes()})
     plotter.draw_edges(color={edge: '#ff0000' for edges in crossings for edge in edges})
 
     plotter.show()