+++ src/compas_rhino/geometry/surface.py
@@ -5,7 +5,7 @@ from __future__ import division
 import compas_rhino
 
 from compas.datastructures import Mesh
-from compas_rhino.geometry import RhinoGeometry
+from compas_rhino.geometry._geometry import RhinoGeometry
 # from compas.geometry import subtract_vectors
 # from compas.geometry import angle_vectors
 # from compas.geometry import distance_point_point